+++ ostap/math/tests/test_math_linalgg.py
@@ -47,6 +47,34 @@ def test_linalg_PLU ( M = 4 , N = 4 ) :
 
     logger.info ( '(P)LU max-difference %.3g :\n%s' % ( delta , D ) ) 
 
+def test_linalg_PQR ( M = 4 , N = 4 ) :
+    
+    logger = getLogger ( 'test_linalg_(P)QR(%s,%s)' % ( M , N )  )
+    
+    A = Matrix ( M , N )
+    for i in range ( A.kRows ) :
+        for j in range ( A.kCols ) :
+            A.set ( i , j , i + j + random.gauss ( 1 , 1 ) )
+            
+    logger.info ( '(P)QR The matrix is:\n%s' % A )
+
+    ## PLU decomposiiton 
+    P, Q, R = A.PQR () 
+    
+    logger.info ( '(P)QR decomposition: P :\n%s' % P )
+    logger.info ( '(P)QR decomposition: Q :\n%s' % Q )
+    logger.info ( '(P)QR decomposition: R :\n%s' % R )
+
+    
+    D     = Q * R - A * P
+    delta = Ostap.Math.maxabs_element ( D ) 
+    
+    logger.info ( '(P)QR max-difference %.3g :\n%s' % ( delta , D ) ) 
+
+    QQ  = Q*Q.T()
+    QQ -= 1 
+    delta2 = Ostap.Math.maxabs_element ( QQ )    
+    logger.info ( '(P)QR non-orthogonality of Q %.3g \n%s' % ( delta2 , QQ ) ) 
     
 # =============================================================================
 if '__main__' == __name__ :
@@ -54,6 +82,10 @@ if '__main__' == __name__ :
     test_linalg_PLU ( 3, 6 )
     test_linalg_PLU ( 3, 3 )
     test_linalg_PLU ( 6, 3 )
+
+    test_linalg_PQR ( 3, 6 )
+    test_linalg_PQR ( 3, 3 )
+    test_linalg_PQR ( 6, 3 )
     
 # =============================================================================
 ##                                                                      The END